Add default layout in the event that local storage saved layout is corrupt
Moving to a typescript enum for LayoutPreset changed how it was encoded (went from object mapping to typescript encoding the enum using integer keys). This threw an error when loading the saved layout.
Just set a default one in case the getter returns null and it should handle any future layout changes.