Fix field flickering (fixes #141)
Pushing default field drawing to the server reintroduced flickering
when multiple packets are published per loop with only one having a
non-empty field overlay. While it seems to the user that only one
field overlay is non-empty, all of the packets with default field
drawing have their own overlays.

The issue is mitigated by putting default field drawing in a separate
canvas and keeping the overlay for explicit, user-visible drawing.